Create a Unix shell-command piping system
This allows seamlessly mixing Python and shell tools. This is a
batteries-included approach, an attempt to add enough convenience
to make Python an appealing replacement for bash shell scripting.

Command exit-codes are checked and handled flexibly, and there is
also support for evaluating a pipeline in the background (forking)
to prevent blocking and support tools like xclip.

The REPL-detection, combining introspection and evalution on repr()
makes the python REPL feel a lot like a shell, making exploration
and experimentation a bit more intuitive and fun.